requirements.txt
fastapi[all]==0.111.1
uvicorn==0.30.1
psycopg2-binary==2.9.9
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
sentence-transformers==3.0.1
torch==2.4.0
tqdm==4.66.4
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.37.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
joblib==1.4.2
scipy==1.10.1
urllib3==2.2.2
pydantic==2.8.2
matplotlib==3.7.5
seaborn==0.13.2
transformers==4.42.4
tmdbsimple==2.9.1